[FIX] helpdesk_mgmt: modify JS in order to make it compatible with website
