+++ helpdesk_mgmt/static/description/index.html
@@ -413,7 +413,7 @@ channels.</li>
 <li>Set the name for the channel.</li>
 <li>You can also Activate or Deactivate channels.</li>
 </ol>
-<p><img alt="image1" src="https://raw.githubusercontent.com/OCA/helpdesk/17.0/helpdesk_mgmt/static/description/Channels.PNG" /></p>
+<p><img alt="image0" src="https://raw.githubusercontent.com/OCA/helpdesk/17.0/helpdesk_mgmt/static/description/Channels.PNG" /></p>
 </div>
 <div class="section" id="categories">
 <h2><a class="toc-backref" href="#toc-entry-3">Categories</a></h2>
@@ -424,7 +424,7 @@ categories.</li>
 <li>Set the name for the category.</li>
 <li>You can also Activate or Deactivate categories.</li>
 </ol>
-<p><img alt="image2" src="https://raw.githubusercontent.com/OCA/helpdesk/17.0/helpdesk_mgmt/static/description/Categories.PNG" /></p>
+<p><img alt="image1" src="https://raw.githubusercontent.com/OCA/helpdesk/17.0/helpdesk_mgmt/static/description/Categories.PNG" /></p>
 </div>
 <div class="section" id="stages">
 <h2><a class="toc-backref" href="#toc-entry-4">Stages</a></h2>
@@ -441,7 +441,7 @@ tickets.</li>
 <li>You can add a description for the stage.</li>
 <li>You can also Activate or Deactivate stages.</li>
 </ol>
-<p><img alt="image3" src="https://raw.githubusercontent.com/OCA/helpdesk/17.0/helpdesk_mgmt/static/description/Stages.PNG" /></p>
+<p><img alt="image2" src="https://raw.githubusercontent.com/OCA/helpdesk/17.0/helpdesk_mgmt/static/description/Stages.PNG" /></p>
 <p>You can also sort the stage sequence if you move up or down the stages
 in the list view.</p>
 </div>
@@ -454,7 +454,7 @@ in the list view.</p>
 <li>Add the teams members.</li>
 <li>You can also Activate or Deactivate teams.</li>
 </ol>
-<p><img alt="image4" src="https://raw.githubusercontent.com/OCA/helpdesk/17.0/helpdesk_mgmt/static/description/Teams.PNG" /></p>
+<p><img alt="image3" src="https://raw.githubusercontent.com/OCA/helpdesk/17.0/helpdesk_mgmt/static/description/Teams.PNG" /></p>
 </div>
 <div class="section" id="tags">
 <h2><a class="toc-backref" href="#toc-entry-6">Tags</a></h2>
@@ -466,7 +466,7 @@ tags.</li>
 <li>Set the color index for the tag.</li>
 <li>You can also Activate or Deactivate tags.</li>
 </ol>
-<p><img alt="image5" src="https://raw.githubusercontent.com/OCA/helpdesk/17.0/helpdesk_mgmt/static/description/Tags.PNG" /></p>
+<p><img alt="image4" src="https://raw.githubusercontent.com/OCA/helpdesk/17.0/helpdesk_mgmt/static/description/Tags.PNG" /></p>
 </div>
 <div class="section" id="permissions">
 <h2><a class="toc-backref" href="#toc-entry-7">Permissions</a></h2>
